+++ app/src/debug/res/values/hyperion.xml
@@ -2,6 +2,4 @@
 <resources>
 
     <string name="hyperion_show_demo">操作デモの表示</string>
-    <string name="hyperion_show_notify_dialog_title">結果通知ダイアログの表示</string>
-    <string name="hyperion_show_retry_dialog_title">リトライダイアログの表示</string>
 </resources>
